refactor(monitor): move @ovenwand/monitor to services
- Renames package to @ovenwand/services.monitor
