+++ .changeset/few-trees-rush.md
@@ -1,0 +1,5 @@
+---
+'@ovenwand/services.monitor': patch
+---
+
+Move package to services and rename to @ovenwand/services.monitor
