+++ apps/monitor.ovenwand.com/package.json
@@ -19,7 +19,7 @@
 	"dependencies": {
 		"@ovenwand/app": "workspace:0.1.17",
 		"@ovenwand/env": "workspace:0.0.2",
-		"@ovenwand/monitor": "workspace:0.2.2",
+		"@ovenwand/services.monitor": "workspace:0.2.2",
 		"@ovenwand/services.faunadb": "workspace:0.2.1",
 		"@ovenwand/ui": "workspace:0.0.17",
 		"@ovenwand/util": "workspace:0.0.17",
