+++ packages/app/package.json
@@ -14,7 +14,7 @@
 	},
 	"dependencies": {
 		"@ovenwand/brand": "workspace:0.0.9",
-		"@ovenwand/monitor": "workspace:0.2.2",
+		"@ovenwand/services.monitor": "workspace:0.2.2",
 		"@ovenwand/ui": "workspace:0.0.17",
 		"@ovenwand/util.browser": "workspace:0.0.12"
 	},
