+++ packages/services/monitor/tsconfig.json
