+++ packages/services/src/index.ts
@@ -1,2 +1,4 @@
 export * from '@ovenwand/services.faunadb';
+export * from '@ovenwand/services.monitor';
+export * from '@ovenwand/services.pusher';
 export * from '@ovenwand/services.sentry';
